+++ src/Nono.Engine/Combinations.cs
@@ -60,14 +60,14 @@ namespace Nono.Engine
 
         /// Also reduces all blocks in cues to length of 1 and space to length of 1.
         /// </summary>
-        public static long Count(ReadOnlySpan<ushort> cues, int length)
+        public static long Count(ReadOnlySpan<int> cues, int length)
         {
             var extraBlockSpace = (MIN_BLOCK_SPACE - 1) * (cues.Length - 1);
             var extraBlockLength = cues.Sum() - cues.Length;
             return CountZippedBlocks(cues.Length, (length - extraBlockSpace - extraBlockLength));
         }
 
-        public static bool IsHot(ReadOnlySpan<ushort> cues, int length)
+        public static bool IsHot(ReadOnlySpan<int> cues, int length)
         {
             if (length < 0) throw new ArgumentOutOfRangeException(nameof(length));
 
@@ -75,7 +75,7 @@ namespace Nono.Engine
             return Moves(cues, length) < maxBlock;
         }
 
-        public static int Moves(ReadOnlySpan<ushort> cues, int length)
+        public static int Moves(ReadOnlySpan<int> cues, int length)
         {
             unchecked
             {
